Move unlauncher_apps to proto package
