+++ app/src/main/java/com/sduduzog/slimlauncher/datasource/apps/UnlauncherAppsSerializer.kt
@@ -3,7 +3,7 @@ package com.sduduzog.slimlauncher.datasource.apps
 import androidx.datastore.core.CorruptionException
 import androidx.datastore.core.Serializer
 import com.google.protobuf.InvalidProtocolBufferException
-import com.jkuester.unlauncher.datastore.UnlauncherApps
+import com.jkuester.unlauncher.datastore.proto.UnlauncherApps
 import java.io.InputStream
 import java.io.OutputStream
