+++ app/src/main/java/com/sduduzog/slimlauncher/ui/main/HomeFragment.kt
@@ -38,7 +38,7 @@ import com.jkuester.unlauncher.datasource.quickbuttonprefs.QuickButtonPreference
 import com.jkuester.unlauncher.datasource.quickbuttonprefs.getIconResourceId
 import com.jkuester.unlauncher.datastore.ClockType
 import com.jkuester.unlauncher.datastore.SearchBarPosition
-import com.jkuester.unlauncher.datastore.UnlauncherApp
+import com.jkuester.unlauncher.datastore.proto.UnlauncherApp
 import com.sduduzog.slimlauncher.R
 import com.sduduzog.slimlauncher.adapters.AppDrawerAdapter
 import com.sduduzog.slimlauncher.adapters.HomeAdapter
