+++ app/src/main/proto/unlauncher_apps.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-option java_package = "com.jkuester.unlauncher.datastore";
+option java_package = "com.jkuester.unlauncher.datastore.proto";
 option java_multiple_files = true;
 
 message UnlauncherApp {
